Pull out chunk based atlas into new file
